Find the longest border.
You are given a table with  rows and columns. Each cell is colored with white or black. Considering the shapes created by black cells, what is the maximum border of these shapes?
A shape is a set of connected cells. Two cells are connected if they share an edge. Note that no shape has a hole in it."
Input:
row = 4
col = 13

.....#####...
......##.....
########.....
...#.........

Output: 6